+++ db/migrate/20170418035928_add_index_on_donations_on_created_at.rb
@@ -1,0 +1,7 @@
+class AddIndexOnDonationsOnCreatedAt < ActiveRecord::Migration
+  disable_ddl_transaction!
+
+  def change
+    add_index :donations, :created_at, algorithm: :concurrently
+  end
+end