+++ db/migrate/20170418040007_add_index_on_master_companies_on_name.rb
@@ -1,0 +1,7 @@
+class AddIndexOnMasterCompaniesOnName < ActiveRecord::Migration
+  disable_ddl_transaction!
+
+  def change
+    add_index :master_companies, :name, algorithm: :concurrently
+  end
+end