+++ spec/acceptance/api/v2/appeals/contacts_spec.rb
@@ -23,6 +23,7 @@ resource 'Appeals > Contacts' do
       church_name
       created_at
       deceased
+      direct_deposit
       donor_accounts
       envelope_greeting
       greeting