+++ spec/acceptance/api/v2/contacts/referrers_spec.rb
@@ -22,6 +22,7 @@ resource 'Contacts > Referrers' do
       church_name
       created_at
       deceased
+      direct_deposit
       donor_accounts
       envelope_greeting
       greeting