+++ spec/acceptance/api/v2/contacts_spec.rb
@@ -59,6 +59,7 @@ resource 'Contacts' do
       church_name
       created_at
       deceased
+      direct_deposit
       donor_accounts
       envelope_greeting
       greeting